+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/dialogs/KingdomCheckDialog.kt
@@ -98,7 +98,6 @@ import kotlin.collections.filter
 import kotlin.collections.map
 import kotlin.collections.plus
 import kotlin.collections.toSet
-import kotlin.collections.toTypedArray
 import kotlin.emptyArray
 import kotlin.js.Promise
 import kotlin.let
@@ -107,9 +106,11 @@ import kotlin.sequences.filter
 import kotlin.sequences.map
 import kotlin.sequences.toSet
 import kotlin.takeIf
+import kotlin.text.contains
 import kotlin.text.isNotBlank
 import kotlin.text.toInt
 import kotlin.to
+import kotlin.toString
 
 @Suppress("unused")
 @JsPlainObject