+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Allow to override hex mile size to 3 or 6 miles in camping settings
+* Add a setting to cap settlement item bonus at kingdom level
 
 ### Fixed